Input & Output
Character Input
Python programming has several in-built functions to perform input and output tasks. Two commonly used functions for I/O (Input/Output) are print and raw_input().. Write a program to get a character as input and print it. The raw_input() function reads formatted input from the keyboard. When the user enters a character, it is stored in variable testCharacter. NOTE: Bold letters or words are Input and rest are output statements.


Sample Input:



z



Sample Output:



z