Operators
Kamal was traveling from Korea to USA and he was not aware of the currency system of USA. Can you please help him to add two dollars and cents?

Note: 1 dollar=100 cents

Input Format

Input consists of 4 integers. The first two inputs correspond to the value of the first dollar and cent. The next two inputs correspond to the value of the second dollar and cent.

Constraints

NA

Output Format

The output should print the sum of dollars and cents.

Sample Input 0

30
10
140
99
Sample Output 0

171
9
Sample Input 1

99
99
99
99
Sample Output 1

199
98