Decision Making
Janu is very interested in mathematics and she loves to solve math puzzles daily. Her friend knows that and she asks her to find out what an automorphic number is and to implement that concept using a programming language. But since Janu is not good at programming, can you help her implement this concept?

An automorphic number is a number whose square ends with the number itself. For example, 5 is an automorphic number as 5^2=25. The last digit is 5 which is the same as the given number. If it is an automorphic number display Automorphic Number else display Not an Automorphic Number.

Input Format

Input consists of an integer

Constraints

NA

Output Format

The output prints one string

Sample Input 0

5
Sample Output 0

Automorphic Number 
Explanation 0

Here, 5^2=25, the last digit of the square value and input is the same and hence it prints the Automorphic Number.

Sample Input 1

7
Sample Output 1

Not an Automorphic Number 
Explanation 1

Here, 7^2=49, the last digit of the square value and input is not the same and hence it prints Not an Automorphic Number.