Strings
Lee conducted a word game for his colleagues. The game is everyone should say a word that should not have any repeating characters in it. If a single character is repeated then the particular person can't continue the game. Lee finds it difficult to disqualify the person from the game since he is not able to find the nonrepetitive character for all the words. Help him to find the winner of the game by writing a program to find the first element which is non -repetitive i.e that element must not be present anywhere else in the string.

Input Format

The first line of the input consists of a string.

Constraints

NA

Output Format

The output displays a character that is non-repetitive. If all the characters in an input string are repetitive, then display All characters are repetitive.

Sample Input 0

teeterson 
Sample Output 0

r
Explanation 0

Here in the word teeterson the first non-repeating character is r and hence it prints r.

Sample Input 1

goog
Sample Output 1

All the characters are repetitive
Explanation 1

Here in the word goog all the characters are repetitive and hence it prints All the characters are repetitive.